Rename: 커스텀 훅 네이밍 컨벤션에 맞게 수정
