+++ src/hooks/api/useThreadsByChannel.ts
@@ -1,15 +1,14 @@
 import { useLocation } from "react-router-dom";
 
-import useChannelDetailsQuery from "@/apis/channel/useChannelDetailsQuery";
-import useThreadsQuery from "@/apis/thread/useGetThreads";
+import useGetChannelDetails from "@/apis/channel/useGetChannelDetails";
+import useGetThreads from "@/apis/thread/useGetThreads";
 
 const useThreadsByChannel = () => {
   const location = useLocation();
   const channelName = location.pathname.split("/").pop() || "compliment";
 
-  // TODO: main merge전 channelName으로 수정해서 올리기 (2024.01.03)
-  const channelQuery = useChannelDetailsQuery(channelName);
-  const threadsQuery = useThreadsQuery(channelQuery.channelDetails?._id);
+  const channelQuery = useGetChannelDetails(channelName);
+  const threadsQuery = useGetThreads(channelQuery.channelDetails?._id);
 
   return {
     channelName,
